- do not use 64bit integer instead of size_t (can't be alloc'ed), or when the actual possible values are 32bit or lower only

+++ sapi/cli/php_http_parser.h
@@ -112,7 +112,7 @@ struct php_http_parser {
   unsigned char index;
 
   uint32_t nread;
-  int64_t content_length;
+  size_t content_length;
 
   /** READ-ONLY **/
   unsigned short http_major;
